+++ packages/react/src/index.tsx
@@ -1,5 +1,8 @@
 import '../index.css';
+// Form
 export * from './components/Button';
+export * from './components/Checkbox';
+export * from './components/Switch';
+// Overlay
 export * from './components/Tooltip';
 export * from './components/Dialog';
-export * from './components/Checkbox';