+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wolves-league-ui/docs
 
+## 0.2.0
+
+### Minor Changes
+
+- Creates `Switch` component story
+
+## 0.1.2
+
+### Patch Changes
+
+- Improves `Checkbox` story
+
 ## 0.1.1
 
 ### Patch Changes